# Use JTS on Spark instead of GT geometries