--- a/src/.sbtopts
+++ b/src/.sbtopts
@@ -1,0 +1,5 @@
+-J-Xmx2g
+-J-XX:+CMSClassUnloadingEnabled
+-J-XX:+UseConcMarkSweepGC
+-Djava.awt.headless=true
+-Dsun.io.serialization.extendedDebugInfo=true