--- a/src/common/build.sbt
+++ b/src/common/build.sbt
@@ -13,6 +13,7 @@ libraryDependencies ++= Seq(
   "com.github.seratch" %% "awscala" % "0.6.1",
   "org.scalaj" %% "scalaj-http" % "2.3.0",
   sparkHive % "provided",
+  sparkJts,
   gtS3 exclude ("com.google.protobuf", "protobuf-java") exclude ("com.amazonaws", "aws-java-sdk-s3"),
   gtSpark exclude ("com.google.protobuf", "protobuf-java"),
   gtVectorTile exclude ("com.google.protobuf", "protobuf-java"),
@@ -32,6 +33,11 @@ libraryDependencies ++= Seq(
   "com.softwaremill.macmemo" %% "macros" % "0.4",
   "com.amazonaws" % "aws-java-sdk-s3" % "1.11.340" % "provided"
 )
+
+Test / fork := true
+Test / baseDirectory := (baseDirectory.value).getParentFile
+Test / parallelExecution := false
+Test / testOptions += Tests.Argument("-oDF")
 
 initialCommands in console :=
   """