--- a/src/common/src/test/scala/common/FeatureBijectionTest.scala
+++ b/src/common/src/test/scala/common/FeatureBijectionTest.scala
@@ -2,7 +2,6 @@ package osmesa.common
 
 import geotrellis.vector._
 import geotrellis.geotools._
-// import geotrellis.geomesa.geotools._
 import org.scalatest._
 import org.geotools.feature.simple.SimpleFeatureImpl