… src/analytics/src/main/scala/osmesa/analytics/oneoffs/StatsJob.scala
… src/common/src/test/resources/isle-of-man-latest.osm.orc
… src/ingest/src/test/scala/osmesa/DebugRelations.scala
… src/ingest/src/test/scala/osmesa/MultiPolygonRelationReconstructionSpec.scala